trying to make prev/next work in ui

+++ src/redux/actions/todos.ts
@@ -1,6 +1,6 @@
 import { ThunkAction as BaseThunkAction } from 'redux-thunk';
 
-import { Todo } from '../../types';
+import { State, Todo } from '../../types';
 
 import db from '../../firebase/firebase';
 
@@ -26,7 +26,7 @@ type DeleteTodo = {
 
 export type Action = AddTodos | CompleteTodo | DeleteTodo;
 
-type ThunkAction = BaseThunkAction<void, {}, {}, Action>;
+type ThunkAction = BaseThunkAction<void, State, {}, Action>;
 
 export function completeTodo(id: string, completed: boolean): ThunkAction {
   return (dispatch) => {
@@ -39,7 +39,21 @@ export function completeTodo(id: string, completed: boolean): ThunkAction {
 }
 
 export function deleteTodo(id: string): ThunkAction {
-  return (dispatch) => {
+  return (dispatch, getState) => {
+    const state: State = getState();
+
+    const batch = db.batch();
+
+    const { prev, next } = state.todos.entities.byId[id];
+    if (prev) {
+      const prevRef = db.collection('todos').doc(prev);
+      batch.update(prevRef, { next });
+    }
+    if (next) {
+      const nextRef = db.collection('todos').doc(next);
+      batch.update(nextRef, { prev });
+    }
+
     db.collection('todos').doc(id).delete().then(() => {
       dispatch({ type: DELETE_TODO, id });
     });
